docker/lockmancyber/foolbox-fgsm/requirements.txt
flask
flask-cors
torch==2.2.2
torchvision==0.17.2
foolbox
pillow